Moved authentication components to subdirectories
