+++ src/components/authentication/new-account/CreateAccount.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 
-import fire from '../../config/Fire';
-import Globals from '../../Globals';
-import * as UI from '../../controls/UI';
+import fire from '../../../config/Fire';
+import Globals from '../../../Globals';
+import * as UI from '../../../controls/UI';
 import './styles/CreateAccount.css';
 
 const minPasswordLength = Globals.constants.minPasswordLength;
