+++ src/components/authentication/new-account/RoleSelection.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+import { withRouter } from 'react-router-dom';
+
+import fire from '../../../config/Fire';
+import * as UI from '../../../controls/UI';
+
+class RoleSelection extends Component {
+
+}
+
+export default withRouter(RoleSelection);
